Added healing circle for activating regen + code cleanup
Option to use only regen circles for regen rather than simply out of range of defences. Would be more realistic in a game setting, where regen that fast HAS to be magic :)

Added some healing music audio to the regen.

Bug Fix: Returning objects to object pool has a delay, but when switching to a new projectiles, the pool is being cleared, but then the return is trying to return a null object. Fix: Added null check.